refactor loader to partner integrations
